extract interfaces for Async classes
