+++ core/cas-server-core-events-configuration-cloud-bus/src/test/java/org/apereo/cas/support/events/listener/CasCloudBusConfigurationEventListenerTests.java
@@ -32,7 +32,7 @@ import static org.junit.jupiter.api.Assertions.*;
 public class CasCloudBusConfigurationEventListenerTests {
     @Autowired
     @Qualifier("casCloudBusConfigurationEventListener")
-    private CasCloudBusConfigurationEventListener casCloudBusConfigurationEventListener;
+    private CasCloudBusConfigurationEventListenerImpl casCloudBusConfigurationEventListener;
 
     @Autowired
     private ConfigurableApplicationContext applicationContext;
