+++ core/cas-server-core-events-configuration/src/main/java/org/apereo/cas/support/events/listener/CasConfigurationEventListener.java
@@ -3,8 +3,6 @@ package org.apereo.cas.support.events.listener;
 import org.apereo.cas.support.events.config.CasConfigurationModifiedEvent;
 import org.apereo.cas.util.spring.CasEventListener;
 import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
 
 /**
  * Interface for {@code CasConfigurationEventListenerImpl} to allow spring {@code @Async} support to use JDK proxy.
@@ -18,8 +16,6 @@ public interface CasConfigurationEventListener extends CasEventListener {
      *
      * @param event the event
      */
-    @EventListener
-    @Async
     void handleRefreshEvent(EnvironmentChangeEvent event);
 
     /**
@@ -27,7 +23,5 @@ public interface CasConfigurationEventListener extends CasEventListener {
      *
      * @param event the event
      */
-    @EventListener
-    @Async
     void handleConfigurationModifiedEvent(CasConfigurationModifiedEvent event);
 }
