+++ core/cas-server-core-services-api/src/main/java/org/apereo/cas/services/RegisteredServicesEventListener.java
@@ -4,8 +4,6 @@ import org.apereo.cas.support.events.service.CasRegisteredServiceExpiredEvent;
 import org.apereo.cas.support.events.service.CasRegisteredServicesRefreshEvent;
 import org.apereo.cas.util.spring.CasEventListener;
 import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
 
 /**
  * Interface for {@code RegisteredServicesEventListenerImpl} to allow spring {@code @Async} support to use JDK proxy.
@@ -19,8 +17,6 @@ public interface RegisteredServicesEventListener extends CasEventListener  {
      *
      * @param event the event
      */
-    @EventListener
-    @Async
     void handleRefreshEvent(CasRegisteredServicesRefreshEvent event);
 
     /**
@@ -28,8 +24,6 @@ public interface RegisteredServicesEventListener extends CasEventListener  {
      *
      * @param event the event
      */
-    @EventListener
-    @Async
     void handleEnvironmentChangeEvent(EnvironmentChangeEvent event);
 
     /**
@@ -37,8 +31,6 @@ public interface RegisteredServicesEventListener extends CasEventListener  {
      *
      * @param event the event
      */
-    @EventListener
-    @Async
     void handleRegisteredServiceExpiredEvent(CasRegisteredServiceExpiredEvent event);
 
 }
