+++ core/cas-server-core-services-registry/src/main/java/org/apereo/cas/services/ServiceRegistryInitializerEventListener.java
@@ -2,51 +2,26 @@ package org.apereo.cas.services;
 
 import org.apereo.cas.support.events.config.CasConfigurationModifiedEvent;
 import org.apereo.cas.util.spring.CasEventListener;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
 
 /**
- * This is {@link ServiceRegistryInitializerEventListener}.
- *
- * @author Misagh Moayyed
- * @since 5.3.0
+ * Interface for {@code ServiceRegistryInitializerEventListenerImpl} to allow spring {@code @Async} support to use JDK proxy.
+ * @author Hal Deadman
+ * @since 6.5.0
  */
-@Slf4j
-@RequiredArgsConstructor
-public class ServiceRegistryInitializerEventListener implements CasEventListener {
-    private final ServiceRegistryInitializer serviceRegistryInitializer;
+public interface ServiceRegistryInitializerEventListener extends CasEventListener {
 
     /**
      * Handle refresh event when issued to this CAS server locally.
      *
      * @param event the event
      */
-    @EventListener
-    @Async
-    public void handleRefreshEvent(final EnvironmentChangeEvent event) {
-        LOGGER.trace("Received event [{}]", event);
-        rebind();
-    }
+    void handleRefreshEvent(EnvironmentChangeEvent event);
 
     /**
      * Handle configuration modified event.
      *
      * @param event the event
      */
-    @EventListener
-    @Async
-    public void handleConfigurationModifiedEvent(final CasConfigurationModifiedEvent event) {
-        if (event.isEligibleForContextRefresh()) {
-            rebind();
-        }
-    }
-
-    private void rebind() {
-        LOGGER.info("Refreshing CAS service registry configuration. Stand by...");
-        serviceRegistryInitializer.initialize();
-    }
+    void handleConfigurationModifiedEvent(CasConfigurationModifiedEvent event);
 }
