+++ core/cas-server-core-services/src/main/java/org/apereo/cas/config/CasServiceRegistryInitializationConfiguration.java
@@ -7,7 +7,7 @@ import org.apereo.cas.services.RegisteredService;
 import org.apereo.cas.services.ServiceRegistry;
 import org.apereo.cas.services.ServiceRegistryExecutionPlanConfigurer;
 import org.apereo.cas.services.ServiceRegistryInitializer;
-import org.apereo.cas.services.ServiceRegistryInitializerEventListener;
+import org.apereo.cas.services.ServiceRegistryInitializerEventListenerImpl;
 import org.apereo.cas.services.ServiceRegistryListener;
 import org.apereo.cas.services.ServicesManager;
 import org.apereo.cas.services.resource.AbstractResourceBasedServiceRegistry;
@@ -76,7 +76,7 @@ public class CasServiceRegistryInitializationConfiguration {
         public CasEventListener serviceRegistryInitializerConfigurationEventListener(
             @Qualifier("serviceRegistryInitializer")
             final ServiceRegistryInitializer serviceRegistryInitializer) {
-            return new ServiceRegistryInitializerEventListener(serviceRegistryInitializer);
+            return new ServiceRegistryInitializerEventListenerImpl(serviceRegistryInitializer);
         }
     }
