+++ support/cas-server-support-service-registry-stream/src/main/java/org/apereo/cas/config/CasServicesStreamingConfiguration.java
@@ -1,7 +1,7 @@
 package org.apereo.cas.config;
 
 import org.apereo.cas.configuration.CasConfigurationProperties;
-import org.apereo.cas.services.CasServicesRegistryStreamingEventListener;
+import org.apereo.cas.services.CasServicesRegistryStreamingEventListenerImpl;
 import org.apereo.cas.services.RegisteredService;
 import org.apereo.cas.services.publisher.CasRegisteredServiceStreamPublisher;
 import org.apereo.cas.services.publisher.DefaultCasRegisteredServiceStreamPublisher;
@@ -39,7 +39,7 @@ public class CasServicesStreamingConfiguration {
         final CasRegisteredServiceStreamPublisher casRegisteredServiceStreamPublisher,
         @Qualifier("casRegisteredServiceStreamPublisherIdentifier")
         final PublisherIdentifier casRegisteredServiceStreamPublisherIdentifier) {
-        return new CasServicesRegistryStreamingEventListener(casRegisteredServiceStreamPublisher,
+        return new CasServicesRegistryStreamingEventListenerImpl(casRegisteredServiceStreamPublisher,
             casRegisteredServiceStreamPublisherIdentifier);
     }
